requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.112.1
gunicorn==23.0.0
h11==0.14.0
idna==3.7
packaging==24.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
pytubefix==6.13.0
requests==2.32.3
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
uvicorn-worker==0.2.0
